+++ src/System.Windows.Forms.Design/tests/UnitTests/System/Windows/Forms/Design/InheritedPropertyDescriptorTests.cs
@@ -1,21 +1,10 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.ComponentModel;
 using System.ComponentModel.Design;
 using System.Drawing;
 
 namespace System.Windows.Forms.Design.Tests;
-internal static class ExtensionHelpers
-{
-    public static InheritedPropertyDescriptor GetInheritedPropertyDescriptor(this Control control, string property)
-    {
-        PropertyDescriptorCollection properties = TypeDescriptor.GetProperties(control);
-        PropertyDescriptor propertyDescriptor = properties.Find(property, false);
-        InheritedPropertyDescriptor inheritedPropertyDescriptor = new(propertyDescriptor, control);
-        return inheritedPropertyDescriptor;
-    }
-}
 
 public class InheritedPropertyDescriptorTests
 {
